[Move] #39  내 지갑 화면 경로 변경
