+++ app/src/main/res/layout/fragment_wallet.xml
@@ -5,7 +5,7 @@
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:background="@color/gray_95"
-    tools:context=".ui.home.WalletFragment">
+    tools:context=".ui.wallet.WalletFragment">
 
     <include
         android:id="@+id/toolbar"
@@ -22,7 +22,7 @@
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:layout_marginHorizontal="20dp"
-        android:layout_marginTop="30dp"
+        android:layout_marginTop="15dp"
         android:background="@drawable/background_gray100_round10"
         android:padding="20dp"
         app:layout_constraintTop_toBottomOf="@+id/toolbar"
@@ -134,9 +134,11 @@
         <androidx.recyclerview.widget.RecyclerView
             android:id="@+id/recyclerView_payment"
             android:layout_width="match_parent"
-            android:layout_height="wrap_content"
+            android:layout_height="0dp"
             android:layout_marginTop="15dp"
+            app:layout_constraintBottom_toBottomOf="parent"
             app:layout_constraintTop_toBottomOf="@+id/divider_payment"
+            app:layout_constraintVertical_bias="0.0"
             tools:layout_editor_absoluteX="16dp" />
     </androidx.constraintlayout.widget.ConstraintLayout>
